+++ AffindaAPI/AffindaAPI/AffindaAPIClient.cs
@@ -895,17 +895,18 @@ namespace Affinda.API
         /// <param name="lowPriority"> Explicitly mark this document as low priority. </param>
         /// <param name="compact"> If true, the returned parse result (assuming `wait` is also true) will be a compact version of the full result. </param>
         /// <param name="deleteAfterParse"> If true, no data will be stored after parsing. Only compatible with requests where wait: True. </param>
+        /// <param name="enableValidationTool"> If true, the document will be viewable in the Affinda Validation Tool. Set to False to optimize parsing speed. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads a document for parsing. When successful, returns an `identifier` in the response for subsequent use with the [/documents/{identifier}](#get-/v3/documents/-identifier-) endpoint to check processing status and retrieve results.&lt;br/&gt;
         /// </remarks>
-        public virtual async Task<Response<Document>> CreateDocumentAsync(Stream file = null, string url = null, string data = null, string collection = null, string workspace = null, string wait = null, string identifier = null, string customIdentifier = null, string fileName = null, string expiryTime = null, string language = null, string rejectDuplicates = null, string regionBias = null, string lowPriority = null, string compact = null, string deleteAfterParse = null, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<Document>> CreateDocumentAsync(Stream file = null, string url = null, string data = null, string collection = null, string workspace = null, string wait = null, string identifier = null, string customIdentifier = null, string fileName = null, string expiryTime = null, string language = null, string rejectDuplicates = null, string regionBias = null, string lowPriority = null, string compact = null, string deleteAfterParse = null, string enableValidationTool = null, CancellationToken cancellationToken = default)
         {
             using var scope = _clientDiagnostics.CreateScope("AffindaAPIClient.CreateDocument");
             scope.Start();
             try
             {
-                return await RestClient.CreateDocumentAsync(file, url, data, collection, workspace, wait, identifier, customIdentifier, fileName, expiryTime, language, rejectDuplicates, regionBias, lowPriority, compact, deleteAfterParse, cancellationToken).ConfigureAwait(false);
+                return await RestClient.CreateDocumentAsync(file, url, data, collection, workspace, wait, identifier, customIdentifier, fileName, expiryTime, language, rejectDuplicates, regionBias, lowPriority, compact, deleteAfterParse, enableValidationTool, cancellationToken).ConfigureAwait(false);
             }
             catch (Exception e)
             {
@@ -931,17 +932,18 @@ namespace Affinda.API
         /// <param name="lowPriority"> Explicitly mark this document as low priority. </param>
         /// <param name="compact"> If true, the returned parse result (assuming `wait` is also true) will be a compact version of the full result. </param>
         /// <param name="deleteAfterParse"> If true, no data will be stored after parsing. Only compatible with requests where wait: True. </param>
+        /// <param name="enableValidationTool"> If true, the document will be viewable in the Affinda Validation Tool. Set to False to optimize parsing speed. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks>
         /// Uploads a document for parsing. When successful, returns an `identifier` in the response for subsequent use with the [/documents/{identifier}](#get-/v3/documents/-identifier-) endpoint to check processing status and retrieve results.&lt;br/&gt;
         /// </remarks>
-        public virtual Response<Document> CreateDocument(Stream file = null, string url = null, string data = null, string collection = null, string workspace = null, string wait = null, string identifier = null, string customIdentifier = null, string fileName = null, string expiryTime = null, string language = null, string rejectDuplicates = null, string regionBias = null, string lowPriority = null, string compact = null, string deleteAfterParse = null, CancellationToken cancellationToken = default)
+        public virtual Response<Document> CreateDocument(Stream file = null, string url = null, string data = null, string collection = null, string workspace = null, string wait = null, string identifier = null, string customIdentifier = null, string fileName = null, string expiryTime = null, string language = null, string rejectDuplicates = null, string regionBias = null, string lowPriority = null, string compact = null, string deleteAfterParse = null, string enableValidationTool = null, CancellationToken cancellationToken = default)
         {
             using var scope = _clientDiagnostics.CreateScope("AffindaAPIClient.CreateDocument");
             scope.Start();
             try
             {
-                return RestClient.CreateDocument(file, url, data, collection, workspace, wait, identifier, customIdentifier, fileName, expiryTime, language, rejectDuplicates, regionBias, lowPriority, compact, deleteAfterParse, cancellationToken);
+                return RestClient.CreateDocument(file, url, data, collection, workspace, wait, identifier, customIdentifier, fileName, expiryTime, language, rejectDuplicates, regionBias, lowPriority, compact, deleteAfterParse, enableValidationTool, cancellationToken);
             }
             catch (Exception e)
             {
@@ -2549,15 +2551,16 @@ namespace Affinda.API
         /// <param name="limit"> The numbers of results to return. </param>
         /// <param name="offset"> The number of documents to skip before starting to collect the result set. </param>
         /// <param name="search"> Search for specific values. </param>
+        /// <param name="annotation"> Filter based on annotation ID. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks> Returns the list of all values in a mapping data source. </remarks>
-        public virtual async Task<Response<Paths1Qr7BnyV3MappingDataSourcesIdentifierValuesGetResponses200ContentApplicationJsonSchema>> ListMappingDataSourceValuesAsync(string identifier, int? limit = null, int? offset = null, string search = null, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<Paths1Qr7BnyV3MappingDataSourcesIdentifierValuesGetResponses200ContentApplicationJsonSchema>> ListMappingDataSourceValuesAsync(string identifier, int? limit = null, int? offset = null, string search = null, int? annotation = null, CancellationToken cancellationToken = default)
         {
             using var scope = _clientDiagnostics.CreateScope("AffindaAPIClient.ListMappingDataSourceValues");
             scope.Start();
             try
             {
-                return await RestClient.ListMappingDataSourceValuesAsync(identifier, limit, offset, search, cancellationToken).ConfigureAwait(false);
+                return await RestClient.ListMappingDataSourceValuesAsync(identifier, limit, offset, search, annotation, cancellationToken).ConfigureAwait(false);
             }
             catch (Exception e)
             {
@@ -2571,15 +2574,16 @@ namespace Affinda.API
         /// <param name="limit"> The numbers of results to return. </param>
         /// <param name="offset"> The number of documents to skip before starting to collect the result set. </param>
         /// <param name="search"> Search for specific values. </param>
+        /// <param name="annotation"> Filter based on annotation ID. </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <remarks> Returns the list of all values in a mapping data source. </remarks>
-        public virtual Response<Paths1Qr7BnyV3MappingDataSourcesIdentifierValuesGetResponses200ContentApplicationJsonSchema> ListMappingDataSourceValues(string identifier, int? limit = null, int? offset = null, string search = null, CancellationToken cancellationToken = default)
+        public virtual Response<Paths1Qr7BnyV3MappingDataSourcesIdentifierValuesGetResponses200ContentApplicationJsonSchema> ListMappingDataSourceValues(string identifier, int? limit = null, int? offset = null, string search = null, int? annotation = null, CancellationToken cancellationToken = default)
         {
             using var scope = _clientDiagnostics.CreateScope("AffindaAPIClient.ListMappingDataSourceValues");
             scope.Start();
             try
             {
-                return RestClient.ListMappingDataSourceValues(identifier, limit, offset, search, cancellationToken);
+                return RestClient.ListMappingDataSourceValues(identifier, limit, offset, search, annotation, cancellationToken);
             }
             catch (Exception e)
             {