+++ AffindaAPI/AffindaAPI/Models/Components1Fe3VqtSchemasInvoicedataPropertiesSupplierfaxAllof1.cs
@@ -11,8 +11,10 @@ namespace Affinda.API.Models
     public partial class Components1Fe3VqtSchemasInvoicedataPropertiesSupplierfaxAllof1
     {
         /// <summary> Initializes a new instance of Components1Fe3VqtSchemasInvoicedataPropertiesSupplierfaxAllof1. </summary>
-        public Components1Fe3VqtSchemasInvoicedataPropertiesSupplierfaxAllof1()
+        /// <param name="raw"></param>
+        public Components1Fe3VqtSchemasInvoicedataPropertiesSupplierfaxAllof1(string raw)
         {
+            Raw = raw;
         }
 
         /// <summary> Initializes a new instance of Components1Fe3VqtSchemasInvoicedataPropertiesSupplierfaxAllof1. </summary>