+++ AffindaAPI/AffindaAPI/Models/Components1Roa72HSchemasInvoicedataPropertiesBankswiftAllof1.cs
@@ -11,8 +11,10 @@ namespace Affinda.API.Models
     public partial class Components1Roa72HSchemasInvoicedataPropertiesBankswiftAllof1
     {
         /// <summary> Initializes a new instance of Components1Roa72HSchemasInvoicedataPropertiesBankswiftAllof1. </summary>
-        public Components1Roa72HSchemasInvoicedataPropertiesBankswiftAllof1()
+        /// <param name="raw"></param>
+        public Components1Roa72HSchemasInvoicedataPropertiesBankswiftAllof1(string raw)
         {
+            Raw = raw;
         }
 
         /// <summary> Initializes a new instance of Components1Roa72HSchemasInvoicedataPropertiesBankswiftAllof1. </summary>