+++ AffindaAPI/AffindaAPI/Models/Components1Vvtu5NSchemasInvoicedataPropertiesPaymentamountpaidAllof1.cs
@@ -11,8 +11,10 @@ namespace Affinda.API.Models
     public partial class Components1Vvtu5NSchemasInvoicedataPropertiesPaymentamountpaidAllof1
     {
         /// <summary> Initializes a new instance of Components1Vvtu5NSchemasInvoicedataPropertiesPaymentamountpaidAllof1. </summary>
-        public Components1Vvtu5NSchemasInvoicedataPropertiesPaymentamountpaidAllof1()
+        /// <param name="raw"></param>
+        public Components1Vvtu5NSchemasInvoicedataPropertiesPaymentamountpaidAllof1(string raw)
         {
+            Raw = raw;
         }
 
         /// <summary> Initializes a new instance of Components1Vvtu5NSchemasInvoicedataPropertiesPaymentamountpaidAllof1. </summary>