+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.21.0] - 2024-04-10
+### Fixed
+- Fix invoiceData annotation required fields to match underling text annotation
+
+### Changed
+- Reduce maximum limit from 300 to 100 and remove default value
+
 ## [5.20.0] - 2024-03-27
 ### Added
 - Add AnnotationBatchUpdate endpoint